chore: segregate client side component for csr and make page ssr by default
