+++ src/common/components/Auth/AuthCard.tsx
@@ -15,7 +15,6 @@ export const AuthCard = <T extends ElementType = typeof defaultElement>({
   children,
   title,
   as,
-  onSubmit,
   ...rest
 }: TitleProps<T>) => {
   const Title = as || defaultElement;
@@ -24,9 +23,7 @@ export const AuthCard = <T extends ElementType = typeof defaultElement>({
       <Title className="text-3xl font-semibold text-text-em-high" {...rest}>
         {title}
       </Title>
-      <form className="flex w-full flex-col gap-6" onSubmit={onSubmit}>
-        {children}
-      </form>
+      {children}
     </div>
   );
 };
